refactored dashboard and dashboard routes, added admin component to handle dashboard routes
