+++ src/components/EditUserForm/EditUserForm.css
@@ -8,7 +8,6 @@
     top: 0;
     bottom: 0;
     padding: 1em;
-
 }
 
 .edit-container{
