+++ src/components/NavBarDash/NavBarDash.js
@@ -18,18 +18,14 @@ export default function DashNavBar(props) {
                     <h1 className="nav-logo--dashboard">Start a <span role="img" aria-label="Fire">🔥</span></h1>
                 </Link>
                 <div className="nav-wrapper--dashboard">
-                    <LoginContext.Consumer>
-                        {user => (
-                            <div className="profile-img-wrapper">
-                                <img className="profile-img" 
-                                    src={(props.user.profile_img_link)? props.user.profile_img_link 
-                                    : "https://i0.wp.com/ahfirstaid.org/wp-content/uploads/2014/07/avatar-placeholder.png?fit=204%2C204"} 
-                                    alt="Profile"
-                                    onClick={() => (user.toggleEditView()) ? "" : props.toggleEditView}
-                                />
-                            </div>
-                        )}
-                    </LoginContext.Consumer>
+                    <div className="profile-img-wrapper">
+                        <img className="profile-img" 
+                            src={(props.user.profile_img_link)? props.user.profile_img_link 
+                            : "https://i0.wp.com/ahfirstaid.org/wp-content/uploads/2014/07/avatar-placeholder.png?fit=204%2C204"} 
+                            alt="Profile"
+                            onClick={() => (props.toggleEditView())}
+                        />
+                    </div>
                     <p className="userName">{(props.user.full_name)}</p>
                     <ul>
                         <li>
